Updates to include the DNA primitive
There are a couple of bits flipped between this register and the value read
out of the JTAG. This changes occasionally between builds. Assuming a meta
stability issue, but unsure where as everything runs on the same clock.

Tried running it off a 20MHz clock and still the same issue.

Will come back to this at a later stage.